requirements.txt
Flask==2.0.3
speedtest-cli==2.1.3
Werkzeug==2.0.3